move control to its own directory
