+++ support/hololens/ServoApp/ServoApp.vcxproj
@@ -151,7 +151,6 @@
     <ClInclude Include="ImmersiveMain.h" />
     <ClInclude Include="ImmersiveView.h" />
     <ClInclude Include="logs.h" />
-    <ClInclude Include="OpenGLES.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="App.h">
       <DependentUpon>App.xaml</DependentUpon>
@@ -159,8 +158,9 @@
     <ClInclude Include="BrowserPage.h">
       <DependentUpon>BrowserPage.xaml</DependentUpon>
     </ClInclude>
-    <ClInclude Include="Servo.h" />
-    <ClInclude Include="ServoControl.h" />
+    <ClInclude Include="ServoControl\OpenGLES.h" />
+    <ClInclude Include="ServoControl\Servo.h" />
+    <ClInclude Include="ServoControl\ServoControl.h" />
   </ItemGroup>
   <ItemGroup>
     <ApplicationDefinition Include="App.xaml">
@@ -298,7 +298,6 @@
     <ClCompile Include="Content\SpinningCubeRenderer.cpp" />
     <ClCompile Include="ImmersiveMain.cpp" />
     <ClCompile Include="ImmersiveView.cpp" />
-    <ClCompile Include="OpenGLES.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
@@ -310,8 +309,9 @@
     </ClCompile>
     <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
     <ClCompile Include="logs.cpp" />
-    <ClCompile Include="Servo.cpp" />
-    <ClCompile Include="ServoControl.cpp" />
+    <ClCompile Include="ServoControl\OpenGLES.cpp" />
+    <ClCompile Include="ServoControl\Servo.cpp" />
+    <ClCompile Include="ServoControl\ServoControl.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Midl Include="App.idl">
@@ -320,7 +320,7 @@
     <Midl Include="BrowserPage.idl">
       <DependentUpon>BrowserPage.xaml</DependentUpon>
     </Midl>
-    <Midl Include="ServoControl.idl" />
+    <Midl Include="ServoControl\ServoControl.idl" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
